vis_portfolio/assignment4-Emilbjacobsen/requirements.txt
keras==2.12.0
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.1
Pillow==9.5.0
scikit_learn==1.2.2
scikit-image==0.20.0
tensorflow==2.12.0
